fix: refactor hook: useActionHandlers
