+++ CODEOWNERS
@@ -47,7 +47,9 @@
 /src/components/Tooltip @amje
 /src/components/theme @resure
 
+/src/hooks/useActionHandlers @ogonkov
 /src/hooks/useOutsideClick @NikitaCG
+/src/hooks/useUniqId @ValeraS
 
 # Allow everyone to update dependencies
 /package.json
