+++ src/components/List/constants.ts
@@ -1,4 +1,4 @@
-import {KeyCode} from '../constants';
+import {KeyCode} from '../../constants';
 
 export const ListQa = {
     ACTIVE_ITEM: 'list-active-item',
