+++ src/components/Pagination/components/PaginationInput/PaginationInput.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import {KeyCode} from '../../../constants';
+import {KeyCode} from '../../../../constants';
 import {TextInput, TextInputProps} from '../../../controls';
 import {blockNew} from '../../../utils/cn';
 import i18n from '../../i18n';
