+++ src/components/index.ts
@@ -62,5 +62,4 @@ export {useOnFocusOutside} from './utils/useOnFocusOutside';
 export * from './utils/interactions';
 export * from './utils/xpath';
 export * from './utils/useFileInput/useFileInput';
-export {useActionHandlers} from './utils/useActionHandlers';
 export {getLayersCount} from './utils/LayerManager';
