+++ src/hooks/index.ts
@@ -1,3 +1,4 @@
+export * from './useActionHandlers';
 export * from './useBodyScrollLock';
 export * from './useOutsideClick';
 export * from './usePortalContainer';
