+++ src/hooks/useActionHandlers/README.md
@@ -1,0 +1,21 @@
+<!--GITHUB_BLOCK-->
+
+# useActionHandlers
+
+<!--/GITHUB_BLOCK-->
+
+```tsx
+import {useActionHandlers} from '@gravity-ui/uikit';
+```
+
+The `useActionHandlers` hook emulates behaviour of system controls, that respond to Enter and Spacebar
+
+## Properties
+
+| Name     | Description                |            Type            | Default |
+| :------- | :------------------------- | :------------------------: | :-----: |
+| callback | Callback for keydown event | `(...args: any[]) => any;` |         |
+
+## Result
+
+Keyboard event handler. `React.KeyboardEventHandler`
