+++ src/hooks/useActionHandlers/index.ts
@@ -1,0 +1,2 @@
+export {useActionHandlers} from './useActionHandlers';
+export type {UseActionHandlersProps, UseActionHandlersResult} from './useActionHandlers';
