+++ src/hooks/useActionHandlers/useActionHandlers.ts
@@ -1,11 +1,13 @@
 import React from 'react';
 
-import {KeyCode} from '../constants';
+import {KeyCode} from '../../constants';
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 type AnyFunction = (...args: any[]) => any;
 
-interface UseActionHandlersResult<T> {
+export type UseActionHandlersProps = AnyFunction;
+
+export interface UseActionHandlersResult<T> {
     onKeyDown: React.KeyboardEventHandler<T>;
 }
 
@@ -14,13 +16,16 @@ interface UseActionHandlersResult<T> {
  * @param callback
  * @return {onKeyDown}
  */
-export function useActionHandlers<T>(callback?: AnyFunction): UseActionHandlersResult<T> {
+export function useActionHandlers<T>(
+    callback?: UseActionHandlersProps,
+): UseActionHandlersResult<T> {
     const onKeyDown = React.useCallback(
         (event: React.KeyboardEvent<T>) => {
             if (
                 callback &&
                 [KeyCode.ENTER, KeyCode.SPACEBAR, KeyCode.SPACEBAR_OLD].includes(event.key)
             ) {
+                // eslint-disable-next-line callback-return
                 callback(event);
             }
         },
